Move non-reusable filter set classes into their respective views
